Split Input.tsx and Select.tsx
