+++ client/src/components/ProductPreview.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import { Form } from "./Form";
-import { FancyInput as Input } from "./Input";
+import Input from "./Input";
 
 type Product = {
   id: string;
